+++ finagle-base-http/src/main/scala/com/twitter/finagle/http/codec/context/HttpContext.scala
@@ -52,7 +52,8 @@ object HttpContext {
     Array[HttpContext](
       HttpDeadline,
       HttpRequeues,
-      HttpBackupRequest
+      HttpBackupRequest,
+      HttpRetry
     )
   }